fix(showcase): resolve dashboard links to real shell host via server-threaded shellUrl
