+++ showcase/shell-dashboard/src/components/feature-grid.tsx
@@ -16,6 +16,7 @@ import { LevelStrip } from "@/components/level-strip";
 import { OverlayColumnHeader } from "@/components/overlay-column-header";
 import { RefDepthHeader, RefDepthCell } from "@/components/ref-depth-column";
 import { buildCellModel } from "@/lib/cell-model";
+import { asParityTier } from "@/lib/page-stats";
 import { getRuntimeConfig } from "@/lib/runtime-config.client";
 import type { CatalogCell } from "@/components/depth-utils";
 import type { Overlay } from "@/lib/overlay-types";
@@ -61,6 +62,7 @@ export function computeColumnTally(
   features: Feature[],
   liveStatus: LiveStatusMap,
   connection: ConnectionStatus = "live",
+  now: number = Date.now(),
 ): { green: number; amber: number; red: number; unknown: boolean } {
   if (connection === "error") {
     return { green: 0, amber: 0, red: 0, unknown: true };
@@ -76,12 +78,16 @@ export function computeColumnTally(
     );
     const isWired = integration.demos.some((d) => d.id === feature.id);
 
-    const model = buildCellModel(liveStatus, {
-      slug: integration.slug,
-      featureId: feature.id,
-      isSupported,
-      isWired,
-    });
+    const model = buildCellModel(
+      liveStatus,
+      {
+        slug: integration.slug,
+        featureId: feature.id,
+        isSupported,
+        isWired,
+      },
+      now,
+    );
 
     if (model.chipColor === "green") green++;
     else if (model.chipColor === "amber") amber++;
@@ -104,6 +110,7 @@ export function computeColumnTallyDetail(
   features: Feature[],
   liveStatus: LiveStatusMap,
   connection: ConnectionStatus,
+  now: number = Date.now(),
 ): TallyDetail {
   if (connection === "error") {
     return { green: [], amber: [], red: [], unknown: true };
@@ -119,22 +126,30 @@ export function computeColumnTallyDetail(
     );
     const isWired = integration.demos.some((d) => d.id === feature.id);
 
-    const model = buildCellModel(liveStatus, {
-      slug: integration.slug,
-      featureId: feature.id,
-      isSupported,
-      isWired,
-    });
+    const model = buildCellModel(
+      liveStatus,
+      {
+        slug: integration.slug,
+        featureId: feature.id,
+        isSupported,
+        isWired,
+      },
+      now,
+    );
 
     // Gray → skip (no data / unsupported / unwired)
     if (model.chipColor === "gray") continue;
 
     // Derive dimension from model: D4/D5/D6 failures are "health" (live
     // round-trip/conversation/parity checks); D3 failures are "e2e" (page-load).
+    // The D6 rung uses the LADDER-GATED `d6Effective` (NOT raw `d6.status`) so
+    // this dimension agrees with the rendered gated D6 badge: when the ladder is
+    // broken below D6, d6Effective collapses to null (no D6-specific failure to
+    // attribute — the real lower-rung failure surfaces through D5/D4 below).
     const dimension: TallyItem["dimension"] =
       (model.d6?.exists &&
-        model.d6.status !== null &&
-        model.d6.status !== "green") ||
+        model.d6Effective !== null &&
+        model.d6Effective !== "green") ||
       (model.d5?.exists &&
         model.d5.status !== null &&
         model.d5.status !== "green") ||
@@ -159,13 +174,45 @@ export function computeColumnTallyDetail(
 }
 
 /**
- * Resolve the shell URL used to build Demo / Code / docs-shell links.
- * Reads from the runtime config injected by the root layout. The
- * sentinel `about:blank#shell-url-missing` is set inside
- * runtime-config.ts when the env var is missing in production —
- * visibly broken links, not silent localhost rendering.
+ * Production sentinel the SERVER runtime-config reader emits when `SHELL_URL`
+ * is unset on the Railway service (`getRuntimeConfig().shellUrl`). It is
+ * truthy, so it would otherwise pass the `serverShellUrl` guard below and get
+ * baked into every anchor as `about:blank#shell-url-missing/integrations/...`.
+ *
+ * Mirrors the SSOT in `shell-dashboard/src/lib/runtime-config.ts`
+ * (`PROD_INVALID_SHELL_URL`) — the same literal the verify-deploy dashboard
+ * guard mirrors. Kept as a local const (the SSOT is a module-private const,
+ * not exported) so the value lives in one shape per consumer with a pointer.
  */
-function resolveShellUrl(): string {
+const PROD_INVALID_SHELL_URL = "about:blank#shell-url-missing";
+
+/**
+ * Resolve the shell URL used to build Demo / Code / docs-shell links.
+ *
+ * Prefers the server-threaded value (passed as `serverShellUrl` from the
+ * server component wrapper that reads `SHELL_URL` at request time). This
+ * is the authoritative source: it is the REAL host during SSR, so anchors
+ * are built correctly in the initial HTML — crawlers and no-JS clients get
+ * working links, and the `https://ssr-placeholder.invalid/` sentinel never
+ * leaks into the rendered DOM.
+ *
+ * EXCEPTION: when `SHELL_URL` is unset on the server, `getRuntimeConfig()`
+ * returns the truthy `about:blank#shell-url-missing` env-unset sentinel. That
+ * is NOT a real host, so we must NOT return it (doing so would bake
+ * `about:blank#shell-url-missing/integrations/...` into anchors). Fall through
+ * to the client config in that case — the verify-deploy dashboard guard fails
+ * the deploy loud when this sentinel ships, so this is a defensive fallback.
+ *
+ * Falls back to the CLIENT runtime config (`window.__SHOWCASE_CONFIG__`)
+ * when no server value was threaded (e.g. a stray client-only caller) or when
+ * the server value is the env-unset sentinel. That client fallback returns the
+ * SSR sentinel during server render, so the server-threaded path is strongly
+ * preferred and is what `page.tsx` wires.
+ */
+function resolveShellUrl(serverShellUrl?: string): string {
+  if (serverShellUrl && serverShellUrl !== PROD_INVALID_SHELL_URL) {
+    return serverShellUrl;
+  }
   return getRuntimeConfig().shellUrl;
 }
 
@@ -192,6 +239,8 @@ interface CategorySectionProps {
   shellUrl: string;
   liveStatus: LiveStatusMap;
   connection: ConnectionStatus;
+  /** Shared frozen reference time — see FeatureGridProps.now. */
+  now: number;
   showRefDepth: boolean;
   refCellsByFeature: Map<string, CatalogCell>;
   categoryColSpan: number;
@@ -205,6 +254,7 @@ const CategorySection = React.memo(
     shellUrl,
     liveStatus,
     connection,
+    now,
     showRefDepth,
     refCellsByFeature,
     categoryColSpan,
@@ -243,13 +293,17 @@ const CategorySection = React.memo(
               ? refCellsByFeature.get(feature.id)
               : undefined;
             const refModel = refCell
-              ? buildCellModel(liveStatus, {
-                  slug: refCell.integration,
-                  featureId: refCell.feature ?? feature.id,
-                  isSupported: refCell.status !== "unsupported",
-                  isWired:
-                    refCell.status === "wired" || refCell.status === "stub",
-                })
+              ? buildCellModel(
+                  liveStatus,
+                  {
+                    slug: refCell.integration,
+                    featureId: refCell.feature ?? feature.id,
+                    isSupported: refCell.status !== "unsupported",
+                    isWired:
+                      refCell.status === "wired" || refCell.status === "stub",
+                  },
+                  now,
+                )
               : undefined;
             return (
               <tr
@@ -355,6 +409,7 @@ const CategorySection = React.memo(
     return (
       prev.liveStatus === next.liveStatus &&
       prev.connection === next.connection &&
+      prev.now === next.now &&
       prev.cat === next.cat &&
       prev.renderCell === next.renderCell &&
       prev.integrations === next.integrations &&
@@ -375,10 +430,27 @@ export interface FeatureGridProps {
   subtitle?: string;
   renderCell: CellRenderer;
   minColWidth?: number;
+  /**
+   * Shell host resolved server-side (request-time `SHELL_URL`) and threaded
+   * down from the server component wrapper. When provided, Demo / Code links
+   * are built with the REAL host during SSR — no `ssr-placeholder.invalid`
+   * sentinel in the HTML, and links work pre-hydration. Falls back to the
+   * client runtime config when omitted.
+   */
+  shellUrl?: string;
   /** Merged live-status map from all subscribed dimensions (lifted to page). */
   liveStatus: LiveStatusMap;
   /** Aggregated SSE connection status (lifted to page). */
   connection: ConnectionStatus;
+  /**
+   * Single frozen reference time shared with the page's cells/stats (computed
+   * once per render in dashboard-page.tsx, re-sampled on live-status change or
+   * the 60s tick). Threaded so the column tallies derive staleness from the
+   * SAME `now` as the cells they summarize — without it each `buildCellModel`
+   * call would default to its own `Date.now()`, and the tally `useMemo` would
+   * never re-evaluate on the tick. Defaults to `Date.now()` for stray callers.
+   */
+  now?: number;
   /** When provided, use overlay-aware column headers and ref-depth column. */
   overlays?: Set<Overlay>;
   /** Catalog data — required when overlays is provided, for ref-depth and parity. */
@@ -392,10 +464,12 @@ export function FeatureGrid({
   minColWidth = 220,
   liveStatus,
   connection,
+  now = Date.now(),
   overlays,
   catalog,
+  shellUrl: serverShellUrl,
 }: FeatureGridProps) {
-  const shellUrl = resolveShellUrl();
+  const shellUrl = resolveShellUrl(serverShellUrl);
   // `getIntegrations()` / `getFeatures()` call `.sort()` / array spread on
   // every invocation, returning a fresh array identity. Memoize once per
   // mount so downstream `useMemo`s keyed on these arrays don't identity-
@@ -412,11 +486,11 @@ export function FeatureGrid({
     for (const integration of integrations) {
       out.set(
         integration.slug,
-        computeColumnTally(integration, features, liveStatus, connection),
+        computeColumnTally(integration, features, liveStatus, connection, now),
       );
     }
     return out;
-  }, [integrations, features, liveStatus, connection]);
+  }, [integrations, features, liveStatus, connection, now]);
 
   // Per-bucket feature lists — mirrors tallies but with TallyItem arrays.
   const tallyDetails = useMemo(() => {
@@ -424,11 +498,17 @@ export function FeatureGrid({
     for (const integration of integrations) {
       out.set(
         integration.slug,
-        computeColumnTallyDetail(integration, features, liveStatus, connection),
+        computeColumnTallyDetail(
+          integration,
+          features,
+          liveStatus,
+          connection,
+          now,
+        ),
       );
     }
     return out;
-  }, [integrations, features, liveStatus, connection]);
+  }, [integrations, features, liveStatus, connection, now]);
 
   // Whether to show the parity ref-depth column
   const showRefDepth = overlays ? overlays.has("parity") : false;
@@ -441,7 +521,20 @@ export function FeatureGrid({
     for (const cell of catalog.cells) {
       if (!seen.has(cell.integration)) {
         seen.add(cell.integration);
-        map.set(cell.integration, cell.parity_tier as ParityTier);
+        // Validate against the SAME PARITY_TIERS guard `computeParityStats`
+        // uses (page-stats.ts) instead of an unchecked `as ParityTier` cast —
+        // an unknown/corrupt tier is skipped (logged loud) rather than seeding
+        // the header map with a bogus tier value.
+        const tier = asParityTier(cell.parity_tier);
+        if (tier === undefined) {
+          console.error(
+            `FeatureGrid parityTierMap: unknown parity_tier ${JSON.stringify(
+              cell.parity_tier,
+            )} for integration ${JSON.stringify(cell.integration)} — skipping`,
+          );
+          continue;
+        }
+        map.set(cell.integration, tier);
       }
     }
     return map;
@@ -633,6 +726,7 @@ export function FeatureGrid({
                 shellUrl={shellUrl}
                 liveStatus={liveStatus}
                 connection={connection}
+                now={now}
                 showRefDepth={showRefDepth}
                 refCellsByFeature={refCellsByFeature}
                 categoryColSpan={categoryColSpan}
